fix: allow asset url function to indicate skipping
If there are multiple web sources, the url function may want to indicate
whether the current asset is available in its source, based on the asset.
While sources are already filtered by asset type, there may be different
properties that determine if a source is applicable.
In our case, we use a special prefix on some projects to indicate a special
source (like a special S3 location). The source will always be the first
source, so we can be sure the prefixed projects are always loaded from the
right source. But if the prefix is missing the url function should be able
to indicate a skip by returning false. Throwing an error is not handled and
results in a failure of the app.