feat: create `addHelper` to add helper w/ priority
The list of helpers in `ScratchStorage` is no longer fixed: now it's
possible to add a new helper using the `addHelper` method. Helpers are
sorted by their relative priority values, which are provided at add
time.

As a follow-up task, `WebHelper` should probably be adjusted to
represent only one endpoint, with create/update/delete features for that
one endpoint. A normal Scratch session might then have one `WebHelper`
for projects and a second for images and sounds. The main reason to do
this would be to concentrate the "try the next place" logic in
`ScratchStorage` instead of having a second version of it in
`WebHelper`.